+++ utils/globals.js
@@ -9,6 +9,16 @@ global.ADDRESS = process.env.ADDRESS || '0.0.0.0';
 global.PORT = process.env.PORT || 3000;
 global.DATA_DIR = process.env.DATA_DIR || process.env.DBLOCATION || localDataDir;
 
+// colors
+global.COLOR_UNSCRATCHED = process.env.COLOR_UNSCRATCHED || '#cca11f';
+global.COLOR_UNSCRATCHED_HOVER = process.env.COLOR_UNSCRATCHED_HOVER || '#8f8f8f';
+global.COLOR_SCRATCHED = process.env.COLOR_SCRATCHED || '#c2c2c2';
+global.COLOR_TEXT = process.env.COLOR_TEXT || '#ffffff';
+global.COLOR_OUTLINES = process.env.COLOR_OUTLINES || '#ffffff';
+global.COLOR_SHARE = process.env.COLOR_SHARE || '#cca11f';
+global.COLOR_SHARE_TEXT = process.env.COLOR_SHARE_TEXT || '#ffffff';
+
+
 // allow LOG_LEVEL to be any case
 if ('LOG_LEVEL' in process.env) {
   // verbosity numbers based on syslog - https://en.wikipedia.org/wiki/Syslog
@@ -52,7 +62,15 @@ if (global.LOG_LEVEL == 'DEBUG') {
   console.debug("  global.PORT:", global.PORT);
   console.debug("  global.DATA_DIR:", global.DATA_DIR);
   console.debug("  global.LOG_LEVEL:", global.LOG_LEVEL);
-  console.debug("  global.ENABLE_SHARE:", global.ENABLE_SHARE);
+  console.debug("  global.ENABLE_SHARE:", global.ENABLE_SHARE, "\n");
+
+  console.debug("  global.COLOR_UNSCRATCHED:", global.COLOR_UNSCRATCHED);
+  console.debug("  global.COLOR_UNSCRATCHED_HOVER:", global.COLOR_UNSCRATCHED_HOVER);
+  console.debug("  global.COLOR_SCRATCHED:", global.COLOR_SCRATCHED);
+  console.debug("  global.COLOR_TEXT:", global.COLOR_TEXT);
+  console.debug("  global.COLOR_OUTLINES:", global.COLOR_OUTLINES);
+  console.debug("  global.COLOR_SHARE:", global.COLOR_SHARE);
+  console.debug("  global.COLOR_SHARE_TEXT:", global.COLOR_SHARE_TEXT);
 
   console.debug("\n");
 }